some ideas are included by the performance is not so good so I am going to try something new
I have already
    1. try the vgg16 backbone but it does not help boost up the performance
    2. switch to another loss which penalize more coordinate error but the
    network behaviors are strange (more inaccurate)
    3. too large to combine two backbone
    4. use the VOC for training (which contains fewer pics with humans)

From 1. and 2. I can conclude that our loss function is not correct, because
even if the loss draw to a very low level when using VGG16, there are more
boxes with non-object showing up in the result, whereas the model with
inception_v1 as backbone has higher loss but more accurate result. That means
that we are not penalizing the non-objectness loss correctly.

I have not tried:
    1. batch norm for w/h output to make them closer to the training set mean.
    2. we are not really having variable loss...(just 2 variants)