Can successfully train now.
The original model is successfully restored. I have added a README file
describing that.

I am trying to provide a mechanism for converting the model to tflite model
but fail. It seems that the "toco" tool has some issues with the batch
normalization. So I am going to remove the batch normalization code and try
again. But note that inception_v1/v2 and resnet_v2 backbones have their internal
batch normalization code and I am not going to remove that. I am only going to
do that for VGG since VGG does not have it.